# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product API",
        "description": "API Documentation for Products, Categories, Statuses, and Auth",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/store-status": {
            "post": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Create a new status",
                "operationId": "bfb54bc9917ddb05e595bc66bf823818",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Status created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/get-status/{id}": {
            "get": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Get status details by ID",
                "operationId": "f408a053c76e88bcb1c26b50855bd762",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status details"
                    },
                    "404": {
                        "description": "Status not found"
                    }
                }
            }
        },
        "/api/v1/update-status/{id}": {
            "put": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Update status",
                "operationId": "13e3054a1f9d07dd34e66175985260b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/delete-status/{id}": {
            "delete": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Delete status",
                "operationId": "c2c8f20fc688cfb3049beca4e17d0d55",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    }
}